+++ test/collectionManagement.test.js
@@ -9,8 +9,7 @@ describe('Good registration is handled well', function() {
     document.body.innerHTML = '<header></header><main></main><footer></footer>'
 
     registrations.OK = {
-      char: '.',
-      context: document.querySelector('main'),
+      match: '.',
       callback: function(character) {
         console.log(`We saw a '${character}'`)
       },
@@ -18,8 +17,7 @@ describe('Good registration is handled well', function() {
     }
 
     registrations.specialKey = {
-      char: 'Backspace',
-      context: document.querySelector('main'),
+      match: 'Backspace',
       callback: function(character) {
         console.log(`We saw a '${character}'`)
       },
@@ -27,8 +25,7 @@ describe('Good registration is handled well', function() {
     }
 
     registrations.array = {
-      char: ['+', '=', '.'],
-      context: document.querySelector('main'),
+      match: ['+', '=', '.'],
       callback: function(character) {
         console.log(`We saw member '${character}'`)
       },
@@ -36,16 +33,15 @@ describe('Good registration is handled well', function() {
     }
 
     registrations.OK2 = {
-      regex: /.+/,
-      context: 'main',
+      match: /.+/,
       callback: function(barcode) {
         console.log(`We saw a '${barcode}'`)
       },
       description: `Let's go and see a barcode`,
     }
 
-    registrations.omittedRegex = {
-      context: document.querySelector('header'),
+    registrations.catchAllRegExp = {
+      match: /^/,
       callback: function(barcode) {
         console.log(`Any barcode could match this, but ${barcode} did.`)
       },
@@ -53,8 +49,7 @@ describe('Good registration is handled well', function() {
     }
 
     registrations.simpleRegex = {
-      regex: /^\w+$/,
-      context: document.querySelector('header'),
+      match: /^\w+$/,
       callback: function(barcode) {
         console.log(
           `Barcodes with regular characters could match this, and ${barcode} did.`,
@@ -64,8 +59,7 @@ describe('Good registration is handled well', function() {
     }
 
     registrations.advancedRegex = {
-      regex: /\w+(_\w+)+/,
-      context: document.querySelector('header'),
+      match: /\w+(_\w+)+/,
       callback: function(barcode) {
         console.log(
           `Barcodes with regular characters in a pattern will match this, and ${barcode} did.`,
@@ -87,21 +81,21 @@ describe('Good registration is handled well', function() {
 
   test('A registration is sanity-checked and found OK', () => {
     const spyConsoleError = jest.spyOn(console, 'error')
-    collectionManagement.registerHotkey(registrations.OK)
+    collectionManagement.register(registrations.OK)
 
     expect(spyConsoleError).not.toHaveBeenCalled()
   })
 
   test('A registration with a special key is sanity-checked and found OK', () => {
     const spyConsoleError = jest.spyOn(console, 'error')
-    collectionManagement.registerHotkey(registrations.specialKey)
+    collectionManagement.register(registrations.specialKey)
 
     expect(spyConsoleError).not.toHaveBeenCalled()
   })
 
   test('A registration with multiple keys is sanity-checked and found OK', () => {
     const spyConsoleError = jest.spyOn(console, 'error')
-    collectionManagement.registerHotkey(registrations.array)
+    collectionManagement.register(registrations.array)
 
     expect(spyConsoleError).not.toHaveBeenCalled()
   })
@@ -109,7 +103,7 @@ describe('Good registration is handled well', function() {
   test(`An alternative registration, with a querySelector as context
     is sanity-checked and found OK`, () => {
     const spyConsoleError2 = jest.spyOn(console, 'error')
-    collectionManagement.registerBarcode(registrations.OK2)
+    collectionManagement.register(registrations.OK2)
 
     expect(spyConsoleError2).not.toHaveBeenCalled()
   })
@@ -117,7 +111,7 @@ describe('Good registration is handled well', function() {
   test(`A registration lacking a description, will not be stored`, () => {
     const spyConsoleStorage = jest.spyOn(dataLockBox, 'store')
     registrations.OK2.description = ''
-    collectionManagement.registerBarcode(registrations.OK2)
+    collectionManagement.register(registrations.OK2)
 
     expect(spyConsoleStorage).not.toHaveBeenCalled()
   })
@@ -125,18 +119,9 @@ describe('Good registration is handled well', function() {
   test(`A registration lacking a callback, typeof function, will not be stored`, () => {
     const spyConsoleStorage = jest.spyOn(dataLockBox, 'store')
     registrations.OK2.callback = []
-    collectionManagement.registerBarcode(registrations.OK2)
+    collectionManagement.register(registrations.OK2)
 
     expect(spyConsoleStorage).not.toHaveBeenCalled()
-  })
-
-  test(`An always-true regex in a barcode registration is used
-    when the regex member is omitted`, () => {
-    let noRegexProps = Object.assign({}, registrations.OK2)
-    delete noRegexProps.regex
-    collectionManagement.registerBarcode(noRegexProps)
-
-    expect(noRegexProps.regex instanceof RegExp).toBe(true)
   })
 
   /* Handle matching entries */
@@ -144,7 +129,7 @@ describe('Good registration is handled well', function() {
   test('A registered callback is returned by entryHandler', () => {
     const entry = '.'
     const spyCallback = jest.spyOn(registrations.OK, 'callback')
-    collectionManagement.registerHotkey(registrations.OK)
+    collectionManagement.register(registrations.OK)
     const handle = collectionManagement.hotkeyHandler(entry)
     handle.callback(entry)
 
@@ -154,7 +139,7 @@ describe('Good registration is handled well', function() {
   test('A with a multicharacter key registered callback is returned by entryHandler', () => {
     const entry = 'Backspace'
     const spyCallback = jest.spyOn(registrations.specialKey, 'callback')
-    collectionManagement.registerHotkey(registrations.specialKey)
+    collectionManagement.register(registrations.specialKey)
     const handle = collectionManagement.hotkeyHandler(entry)
     handle.callback(entry)
 
@@ -164,7 +149,7 @@ describe('Good registration is handled well', function() {
   test('A with multiple keys registered callback is returned by entryHandler', () => {
     const entry = '='
     const spyCallback = jest.spyOn(registrations.array, 'callback')
-    collectionManagement.registerHotkey(registrations.array)
+    collectionManagement.register(registrations.array)
     const handle = collectionManagement.hotkeyHandler(entry)
     handle.callback(entry)
 
@@ -176,10 +161,10 @@ describe('Good registration is handled well', function() {
     const entry = '.'
     const entry3 = '+'
 
-    collectionManagement.registerHotkey(registrations.array)
+    collectionManagement.register(registrations.array)
     const handle = collectionManagement.hotkeyHandler(entry)
 
-    collectionManagement.registerHotkey(registrations.OK)
+    collectionManagement.register(registrations.OK)
     const handle2 = collectionManagement.hotkeyHandler(entry)
     const handle3 = collectionManagement.hotkeyHandler(entry3)
 
@@ -192,10 +177,10 @@ describe('Good registration is handled well', function() {
     const entry = '.'
     const entry3 = '+'
 
-    collectionManagement.registerHotkey(registrations.OK)
+    collectionManagement.register(registrations.OK)
     const handle = collectionManagement.hotkeyHandler(entry)
 
-    collectionManagement.registerHotkey(registrations.array)
+    collectionManagement.register(registrations.array)
     const handle2 = collectionManagement.hotkeyHandler(entry)
     const handle3 = collectionManagement.hotkeyHandler(entry3)
 
@@ -206,16 +191,15 @@ describe('Good registration is handled well', function() {
   /* Utiliity functions */
 
   test('All registrations can be reset', () => {
-    collectionManagement.registerHotkey(registrations.OK)
-    collectionManagement.registerBarcode(registrations.OK2)
-    const handle = collectionManagement.hotkeyHandler(registrations.OK.char)
+    const entry = registrations.OK.match
+    collectionManagement.register(registrations.OK)
+    collectionManagement.register(registrations.OK2)
+    const handle = collectionManagement.hotkeyHandler(entry)
     const handle2 = collectionManagement.barcodeHandler('PICTURAE')
 
     collectionManagement.reset()
 
-    const handle_reset = collectionManagement.hotkeyHandler(
-      registrations.OK.char,
-    )
+    const handle_reset = collectionManagement.hotkeyHandler(entry)
     const handle2_reset = collectionManagement.barcodeHandler('PICTURAE')
 
     expect(handle).toBeTruthy()
@@ -227,21 +211,25 @@ describe('Good registration is handled well', function() {
   /* Barcode matching preference */
 
   test(`A barcode is paired to the lengthiest matching Regexp`, () => {
-    const entryOmit = '32457+5456'
-    const spyCallbackOmit = jest.spyOn(registrations.omittedRegex, 'callback')
-    collectionManagement.registerBarcode(registrations.omittedRegex)
+    const entryAny = 'x32457+5456'
+    const spyCallbackAny = jest.spyOn(registrations.catchAllRegExp, 'callback')
+    collectionManagement.register(registrations.catchAllRegExp)
     const entrySimple = 'HEERHUGOWAARD'
     const spyCallbackSimple = jest.spyOn(registrations.simpleRegex, 'callback')
-    collectionManagement.registerBarcode(registrations.simpleRegex)
+    collectionManagement.register(registrations.simpleRegex)
     const entryAdv = 'RESULT_NOT_OK'
     const spyCallbackAdv = jest.spyOn(registrations.advancedRegex, 'callback')
-    collectionManagement.registerBarcode(registrations.advancedRegex)
+    collectionManagement.register(registrations.advancedRegex)
 
-    collectionManagement.barcodeHandler(entryOmit).callback(entryOmit)
-    collectionManagement.barcodeHandler(entrySimple).callback(entrySimple)
-    collectionManagement.barcodeHandler(entryAdv).callback(entryAdv)
+    const handleAny = collectionManagement.barcodeHandler(entryAny)
+    const handleSimple = collectionManagement.barcodeHandler(entrySimple)
+    const handleAdv = collectionManagement.barcodeHandler(entryAdv)
 
-    expect(spyCallbackOmit).toHaveBeenCalledWith(entryOmit)
+    handleAny.callback(entryAny)
+    handleSimple.callback(entrySimple)
+    handleAdv.callback(entryAdv)
+
+    expect(spyCallbackAny).toHaveBeenCalledWith(entryAny)
     expect(spyCallbackSimple).toHaveBeenCalledWith(entrySimple)
     expect(spyCallbackAdv).toHaveBeenCalledWith(entryAdv)
   })
@@ -257,24 +245,24 @@ describe('Good registration is handled well', function() {
   })
 
   test("An overview object with an array 'hotkey' is returned when some hotkeys were configured", () => {
-    collectionManagement.registerHotkey(registrations.OK)
+    collectionManagement.register(registrations.OK)
     const ovvObj = collectionManagement.overviewJson()
 
     expect(ovvObj.hotkeys instanceof Array).toBe(true)
   })
 
   test("An overview object with an array 'hotkey' and an array 'barcode' is returned when some hotkeys and barcodes were configured", () => {
-    collectionManagement.registerHotkey(registrations.OK)
-    collectionManagement.registerBarcode(registrations.OK2)
+    collectionManagement.register(registrations.OK)
+    collectionManagement.register(registrations.OK2)
     const ovvObj = collectionManagement.overviewJson()
 
     expect(ovvObj.barcodes instanceof Array).toBe(true)
   })
 
   test('An help-screen with all active matches is coerced to the hosting application', () => {
-    collectionManagement.registerHotkey(registrations.OK)
-    collectionManagement.registerBarcode(registrations.OK2)
-    collectionManagement.registerBarcode(registrations.omittedRegex)
+    collectionManagement.register(registrations.OK)
+    collectionManagement.register(registrations.OK2)
+    collectionManagement.register(registrations.catchAllRegExp)
     collectionManagement.overviewPanel()
 
     const totalSelector = 'chars-and-chuncks-panel tbody tr'
@@ -289,21 +277,21 @@ describe('Good registration is handled well', function() {
   /* Cleanup */
 
   test('A cleanup function is returned when a hotkey or barcode is registered', () => {
-    const cleanupOK = collectionManagement.registerHotkey(registrations.OK)
+    const cleanupOK = collectionManagement.register(registrations.OK)
 
     expect(typeof cleanupOK === 'function').toBe(true)
   })
 
   test(`A cleanup removes references to the context of a hotkey or barcode,
       and therefore the hotkey or barcode can not be used anymore`, () => {
-    const cleanup = collectionManagement.registerHotkey(registrations.OK)
-    const handle1 = collectionManagement.hotkeyHandler(registrations.OK.char)
+    const entry = registrations.OK.match
+    const cleanup = collectionManagement.register(registrations.OK)
+    const handle1 = collectionManagement.hotkeyHandler(entry)
 
     expect(registrations.OK.box).toBe(handle1)
 
     cleanup()
-    const handle2 = collectionManagement.hotkeyHandler(registrations.OK.char)
-
+    const handle2 = collectionManagement.hotkeyHandler(entry)
     expect(registrations.OK.box).not.toBe(handle2)
   })
 
@@ -311,13 +299,13 @@ describe('Good registration is handled well', function() {
       and so the cleanup can also be done in one call`, () => {
     const spyCleanup = jest.spyOn(dataLockBox, 'cleanup')
 
-    const cleanupRefs = collectionManagement.registerHotkeys([
+    const cleanupRefs = collectionManagement.register([
       registrations.OK,
       registrations.specialKey,
       registrations.array,
     ])
     cleanupRefs()
 
-    expect(spyCleanup).toHaveBeenCalledTimes(3)
+    expect(spyCleanup).toHaveBeenCalledTimes(1)
   })
 })